+++ nodemanager/Sensor.h
@@ -33,46 +33,46 @@ Sensor: provide functionalities common to all the sensors
 class Sensor {
 public:
 	Sensor();
-	Sensor(int pin = -1);
+	Sensor(int8_t pin = -1);
 	// return the name of the sensor
 	const char* getName();
 	// [1] where the sensor is attached to (default: not set)
-	void setPin(int value);
+	void setPin(int8_t value);
 	// [5] For some sensors, the measurement can be queried multiple times and an average is returned (default: 1)
-	void setSamples(int value);
+	void setSamples(unsigned int value);
 	// [6] If more then one sample has to be taken, set the interval in milliseconds between measurements (default: 0)
-	void setSamplesInterval(int value);
+	void setSamplesInterval(unsigned long value);
     // [17] After how many seconds the sensor will report back its measure (default: 10 minutes)
-    void setReportIntervalSeconds(int value);
+    void setReportIntervalSeconds(unsigned long value);
     // [16] After how many minutes the sensor will report back its measure (default: 10 minutes)
-    void setReportIntervalMinutes(int value);
+    void setReportIntervalMinutes(unsigned long value);
     // [19] After how many hours the sensor will report back its measure (default: 10 minutes)
-    void setReportIntervalHours(int value);
+    void setReportIntervalHours(unsigned int value);
     // [20] After how many days the sensor will report back its measure (default: 10 minutes)
-    void setReportIntervalDays(int value);
+    void setReportIntervalDays(uint8_t value);
 	// [24] Set the way the timer used for reporting to the gateway should operate. It can be either TIME_INTERVAL (e.g. report every X seconds with the amount of time set with setReportTimerValue()), IMMEDIATELY (e.g. report at every cycle, useful for sensors like actuators which should report as soon as the value has changed), DO_NOT_REPORT (e.g. never report, useful for when there is no need to report, like a Display) and when NODEMANAGER_TIME is ON, EVERY_MINUTE/EVERY_HOUR/EVERY_DAY (e.g. to report the value set in the previous timeframe, useful for sensors reporting an accumulated value linked to a timeframe at regular intervals), AT_MINUTE/AT_HOUR/AT_DAY (e.g. report at a given minute/hour/day, useful if the measure is expected at a specified time, set with setReportTimerValue())
 	void setReportTimerMode(timer_mode value);
 	// [25] Set the value for the reporting timer's mode which has been set with setReportTimerMode()
-	void setReportTimerValue(int value);
+	void setReportTimerValue(unsigned long value);
 	// [26] Set the way the timer used for taking measures should operate. Takes the same parameters as setReportTimerMode(). If not set explicitly, will be set as the reporting timer
 	void setMeasureTimerMode(timer_mode value);
 	// [27] Set the value for the reporting timer's mode which has been set with setReportTimerMode() If not set explicitely, will be set with the same value as the reporting timer
-	void setMeasureTimerValue(int value);
+	void setMeasureTimerValue(unsigned long value);
 	// list of configured child
 	List<Child*> children;
 	// return the child object based on the provided child_id
-	Child* getChild(int child_id);
+	Child* getChild(uint8_t child_id);
 	// register a child
 	void registerChild(Child* child);
 #if NODEMANAGER_INTERRUPTS == ON
 	// return the pin the interrupt is attached to
-	int getInterruptPin();
+	int8_t getInterruptPin();
 	// set initial value of the configured pin. Can be used for internal pull up
-	void setPinInitialValue(int value);
+	void setPinInitialValue(int8_t value);
 	// for interrupt-based sensor, set the interrupt mode. Can be CHANGE, RISING, FALLING (default: CHANGE)
-	void setInterruptMode(int value);
+	void setInterruptMode(uint8_t value);
 	// [22] for interrupt-based sensor, milliseconds to wait/sleep after the interrupt before reporting (default: 0)
-	void setWaitAfterInterrupt(int value);
+	void setWaitAfterInterrupt(unsigned long value);
 	// [23] for interrupt-based sensor, the value of the pin is checked and the interrupt ignored if RISING and not HIGH or FALLING and not LOW (default: true)
 	void setInterruptStrict(bool value);
 #endif
@@ -80,7 +80,7 @@ public:
 	// set a previously configured PowerManager to the sensor so to powering it up with custom pins
 	void setPowerManager(PowerManager& powerManager);
 	// to save battery the sensor can be optionally connected to two pins which will act as vcc and ground and activated on demand
-	void setPowerPins(int ground_pin, int vcc_pin, int wait_time = 50);
+	void setPowerPins(int8_t ground_pin, int8_t vcc_pin, unsigned long wait_time = 50);
 	// [13] manually turn the power on
 	void powerOn();
 	// [14] manually turn the power off
@@ -118,18 +118,18 @@ public:
 #endif
 protected:
 	const char* _name = "";
-	int _pin = -1;
-	int _samples = 1;
-	int _samples_interval = 0;
+	int8_t _pin = -1;
+	unsigned int _samples = 1;
+	unsigned long _samples_interval = 0;
 	bool _first_run = true;
 	Timer* _report_timer;
 	Timer* _measure_timer;
 	bool _evaluateTimer(Timer* timer);
 #if NODEMANAGER_INTERRUPTS == ON
-	int _interrupt_pin = -1;
-	int _interrupt_mode = MODE_NOT_DEFINED;
-	int _wait_after_interrupt = 0;
-	int _initial_value = -1;
+	int8_t _interrupt_pin = -1;
+	uint8_t _interrupt_mode = MODE_NOT_DEFINED;
+	unsigned long _wait_after_interrupt = 0;
+	int8_t _initial_value = -1;
 	bool _interrupt_strict = true;
 #endif
 #if NODEMANAGER_POWER_MANAGER == ON
